Modify item regex to allow paragraph breaks
Before this commit, a list of the form

```
<li>[ ] something</li>
<li>[ ] something else<p>A paragraph break inside</p>
</li>
<li>[ ] regular something</li>
```

renders incorrectly as there is a line break before `</li>`. This can be
output from using the `mdx_truly_sane_lists` extension.

NOTE:

I note that if one doesn't use the truly sane list extension, then
multiparagraph items in a list element don't render correctly before or
after this commit. Without the truly sane list extension, the string

```
- [ ] one
- [ ] two, multiline

  a second para

- [ ] three
```

is converted internally into

```
<li>[ ] one</li>
<li>
<p>[ ] two, multiline</p>
<p>a second para</p>
</li>
<li>[ ] three</li>
```

which is is not detected by the markdown-checklist postprocessor regex
either. I don't handle that in this commit.